Ch. 12: Make home page scrollable.
Add scrollable content to the home page, and also make the
toolbar scrollable, so that when the page is scrolled down, the
toolbar gets hidden as well. The tab bar doesn't scroll, but
it can be made to scroll by simply adding the same
layout_scrollFlags to the TabLayout as in the Toolbar in
activity_main.xml.